Vex v0.6.0 Vex.ErrorRenderer behaviour

Implementation of this behaviour should be set in validator options as :error_renderer or in :vex application config with same key.

Common pattern is to expect :message key in validator options and if it not set use default_message (use get_message function for it).

Result of message function appears in Vex.errors error tuple as last element, for those validators who use Vex.Validator.ErrorMessage (in general they should).

Functions

get_message(options, default)

Callbacks

message(validator_options, default_message, context)

message(validator_options :: list(), default_message :: String.t(), context :: list()) :: any()